refactor: move WorkflowBuilder to same directory, protect constructors
